Pipe to communicate between Server and Main Thread
To successfully run the server enter Gameplay Segment: Connect Client then provide a sender and receiver ID in 12:15 format. At this point there should be a return on the main terminal. Hit S once and the score on the screen will update based on the server information. <WARNING> If you attempt to update the score multiple times or without first entering a score from the client the Main Thread will wait and freeze until a value is sent from the client.